# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solar Panel ERP API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/production/boms/{bom_id}/explode": {
      "get": {
        "tags": [
          "Production"
        ],
        "summary": "Explode BOM recipe",
        "description": "Pure domain function that explodes a BOM recipe for a target production quantity.\n\n- Formula: round(production_quantity * component_quantity / bom_output_quantity, 6).\n- Read-only: zero database mutations.",
        "operationId": "get_bom_explosion_production_boms__bom_id__explode_get",
        "parameters": [
          {
            "name": "bom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Bom Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                }
              ],
              "description": "Target production quantity",
              "title": "Quantity"
            },
            "description": "Target production quantity"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMExplosionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/boms/{bom_id}/feasibility": {
      "get": {
        "tags": [
          "Production"
        ],
        "summary": "Analyze BOM material feasibility",
        "description": "Evaluates whether sufficient raw components exist in reservable stock to manufacture the requested quantity.\n\n- Checks only active WAREHOUSE, BIN, and PRODUCTION locations.\n- Excludes quarantine, damaged, and transit inventory.\n- Reports exact shortage quantities per component and overall can_produce flag.",
        "operationId": "get_bom_feasibility_production_boms__bom_id__feasibility_get",
        "parameters": [
          {
            "name": "bom_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Bom Id"
            }
          },
          {
            "name": "quantity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,19}0*$)\\d{0,12}\\.\\d{0,6}0*$)"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target production quantity",
              "title": "Quantity"
            },
            "description": "Target production quantity"
          },
          {
            "name": "production_required",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,12}|(?=[\\d.]{1,19}0*$)\\d{0,12}\\.\\d{0,6}0*$)"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Target production quantity (alias)",
              "title": "Production Required"
            },
            "description": "Target production quantity (alias)"
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialFeasibilityRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders": {
      "post": {
        "tags": [
          "Production"
        ],
        "summary": "Create production order",
        "description": "Creates a new production order in DRAFT status.\n\n- Freezes the active BOM version and calculates material requirements.\n- Does not reserve stock; keeps creation and commitment separate.",
        "operationId": "post_production_order_production_orders_post",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionOrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders/{order_id}": {
      "get": {
        "tags": [
          "Production"
        ],
        "summary": "Get production order details",
        "description": "Retrieves production order details, status, and frozen material requirements.",
        "operationId": "get_production_order_production_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders/{order_id}/release": {
      "post": {
        "tags": [
          "Production"
        ],
        "summary": "Release production order and reserve materials",
        "description": "Releases a DRAFT production order (status -> RELEASED) and reserves required component materials.\n\n- Option A Ownership: reservations reference production_order_material_id.\n- All-or-Nothing: fails with 409 Conflict (insufficient_stock) if any component is unavailable.",
        "operationId": "post_release_production_order_production_orders__order_id__release_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ProductionOrderRelease"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders/{order_id}/start": {
      "post": {
        "tags": [
          "Production"
        ],
        "summary": "Start production order",
        "description": "Transitions a RELEASED production order to IN_PROGRESS.\n\n- Idempotent; rejects unreleased or cancelled orders.",
        "operationId": "post_start_production_order_production_orders__order_id__start_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders/{order_id}/complete": {
      "post": {
        "tags": [
          "Production"
        ],
        "summary": "Atomically complete production order",
        "description": "Atomically completes production execution within a single database transaction:\n\n1. Consumes active material reservations (status -> CONSUMED).\n2. Posts raw material consumption movements (PRODUCTION_CONSUMPTION).\n3. Posts finished goods output movement (PRODUCTION_OUTPUT) into destination warehouse.\n4. Marks order COMPLETED.\n- All-or-nothing: if physical balances are insufficient, entire transaction rolls back cleanly.\n- Idempotent: repeated completion of an already completed order returns safely.",
        "operationId": "post_complete_production_order_production_orders__order_id__complete_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ProductionOrderComplete"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "Production"
        ],
        "summary": "Cancel production order",
        "description": "Cancels a production order (status -> CANCELLED) and releases active component reservations.",
        "operationId": "post_cancel_production_order_production_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production-orders": {
      "post": {
        "tags": [
          "Production orders"
        ],
        "summary": "Create production order",
        "description": "Creates a new production order in DRAFT status with frozen material requirements.",
        "operationId": "post_direct_production_order_production_orders_post",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductionOrderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production-orders/{order_id}": {
      "get": {
        "tags": [
          "Production orders"
        ],
        "summary": "Get production order details",
        "description": "Retrieves production order details, current status, and component requirements.",
        "operationId": "get_direct_production_order_production_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/production-orders/{order_id}/release": {
      "post": {
        "tags": [
          "Production orders"
        ],
        "summary": "Release production order and reserve materials",
        "description": "Releases a DRAFT production order and reserves all required component materials.",
        "operationId": "post_direct_release_production_order_production_orders__order_id__release_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "Order Id"
            }
          },
          {
            "name": "x-organization-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 9223372036854775807,
              "exclusiveMinimum": 0,
              "title": "X-Organization-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ProductionOrderRelease"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionOrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BOMExplosionRead": {
        "properties": {
          "bom_id": {
            "type": "integer",
            "title": "Bom Id"
          },
          "production_quantity": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Production Quantity"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/ExplodedRequirementRead"
            },
            "type": "array",
            "title": "Requirements"
          }
        },
        "type": "object",
        "required": [
          "bom_id",
          "production_quantity",
          "requirements"
        ],
        "title": "BOMExplosionRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialFeasibilityRead": {
        "properties": {
          "bom_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bom Id"
          },
          "production_required": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Production Required"
          },
          "can_produce": {
            "type": "boolean",
            "title": "Can Produce"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialFeasibilityItem"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "required": [
          "production_required",
          "can_produce",
          "materials"
        ],
        "title": "MaterialFeasibilityRead"
      },
...
      "ProductionOrderRelease": {
        "properties": {
          "location_id": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id",
            "description": "Optional preferred stock location to reserve components from."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProductionOrderRelease"
      },
...
    }
  },
...
}